requirements.txt
2to3==1.0
alembic==1.5.8
apiron==5.1.0
Babel==2.9.0
barcode==1.0.1
bcrypt==3.2.0
beren==0.7.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.1.7
config==0.5.0.post0
crypto==1.4.1
cycler==0.10.0
Cython==3.0.2
defusedxml==0.7.1
dnspython==2.1.0
email-validator==1.1.2
Flask==2.3.3
Flask-Babel==2.0.0
Flask-Cors==3.0.10
Flask-Env==2.0.0
Flask-Ext==0.1
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
flask-tryton==0.11.1
Flask-WTF==1.1.1
Genshi==0.7.5
greenlet==2.0.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.7.3
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.3.1
lxml==4.9.3
Mako==1.1.4
MarkupSafe==2.1.3
matplotlib==3.3.4
Naked==0.1.31
numpy==1.19.5
passlib==1.7.4
pendulum==2.1.2
Pillow==8.4.0
polib==1.1.0
pycparser==2.20
pycrypto==2.6.1
pydantic==2.6.1
pyparsing==2.4.7
pypng==0.20220715.0
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.20.0
python-editor==1.0.4
python-Levenshtein==0.12.0
python-magic==0.4.22
python-sql==1.2.1
python-stdnum==1.16
pytz==2021.1
pytzdata==2020.1
PyYAML==6.0.1
relatorio==0.9.2
requests==2.25.1
shellescape==3.8.1
simpleeval==0.9.13
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.3.23
urllib3==1.26.4
wrapt==1.12.1
WTForms==2.3.3
zipp==3.4.1
